Selection of user skills
Depending on the programming skill of the user, some element types can be hidden. So a beginner will not see element subtypes which he never will use.
User can select whether he is a beginner, advanced or programmer in the settings.

Warnings are shown if user wants to edit or save a demonstration flow. A demonstration flow can be copied and it becomes a normal flow.